+++ internal/pkg/instrumentation/bpf/go.opentelemetry.io/otel/traceglobal/probe.go
@@ -6,6 +6,7 @@ package global
 import (
 	"encoding/binary"
 	"fmt"
+	"log/slog"
 	"math"
 
 	"go.opentelemetry.io/auto/internal/pkg/inject"
@@ -13,7 +14,6 @@ import (
 	"go.opentelemetry.io/auto/internal/pkg/process"
 	"go.opentelemetry.io/auto/internal/pkg/structfield"
 
-	"github.com/go-logr/logr"
 	"github.com/hashicorp/go-version"
 	"golang.org/x/sys/unix"
 
@@ -32,14 +32,14 @@ const (
 )
 
 // New returns a new [probe.Probe].
-func New(logger logr.Logger) probe.Probe {
+func New(logger *slog.Logger) probe.Probe {
 	id := probe.ID{
 		SpanKind:        trace.SpanKindClient,
 		InstrumentedPkg: pkg,
 	}
 	return &probe.Base[bpfObjects, event]{
 		ID:     id,
-		Logger: logger.WithName(id.String()),
+		Logger: logger,
 		Consts: []probe.Const{
 			probe.RegistersABIConst{},
 			probe.AllocationConst{},