+++ internal/tools/inspect/cmd/offsetgen/main.go
@@ -10,13 +10,10 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"log"
+	"log/slog"
 	"os"
 	"os/signal"
 	"runtime"
-
-	"github.com/go-logr/logr"
-	"github.com/go-logr/stdr"
 
 	"go.opentelemetry.io/auto/internal/pkg/structfield"
 	"go.opentelemetry.io/auto/internal/tools/inspect"
@@ -38,7 +35,7 @@ var (
 	// verbosity is the log verbosity level flag value.
 	verbosity int
 
-	logger logr.Logger
+	logger *slog.Logger
 )
 
 func init() {
@@ -49,8 +46,7 @@ func init() {
 
 	flag.Parse()
 
-	stdr.SetVerbosity(verbosity)
-	logger = stdr.New(log.New(os.Stderr, "", log.LstdFlags))
+	logger = slog.Default()
 }
 
 func manifests() ([]inspect.Manifest, error) {
@@ -199,7 +195,7 @@ func main() {
 func run() error {
 	m, err := manifests()
 	if err != nil {
-		logger.Error(err, "failed to load manifests")
+		logger.Error("failed to load manifests", "error", err)
 		return err
 	}
 
@@ -207,14 +203,14 @@ func run() error {
 	if cacheFile != "" {
 		cache, err = inspect.NewCache(logger, cacheFile)
 		if err != nil {
-			logger.Error(err, "failed to load cache", "path", cacheFile)
+			logger.Error("failed to load cache", "error", err, "path", cacheFile)
 			// Use an empty cache.
 		}
 	}
 
 	i, err := inspect.New(logger, cache, m...)
 	if err != nil {
-		logger.Error(err, "failed to setup inspector")
+		logger.Error("failed to setup inspector", "error", err)
 		return err
 	}
 	i.NWorkers = numCPU
@@ -225,7 +221,7 @@ func run() error {
 
 	to, err := i.Do(ctx)
 	if err != nil {
-		logger.Error(err, "failed get offsets")
+		logger.Error("failed get offsets", "error", err)
 		return err
 	}
 
@@ -237,7 +233,7 @@ func run() error {
 	logger.Info("writing offsets", "dest", outputFile)
 	f, err := os.Create(outputFile)
 	if err != nil {
-		logger.Error(err, "failed to open output file", "dest", outputFile)
+		logger.Error("failed to open output file", "error", err, "dest", outputFile)
 		return err
 	}
 	defer f.Close()
@@ -245,7 +241,7 @@ func run() error {
 	enc := json.NewEncoder(f)
 	enc.SetIndent("", "  ")
 	if err := enc.Encode(to); err != nil {
-		logger.Error(err, "failed to write offsets", "dest", outputFile)
+		logger.Error("failed to write offsets", "error", err, "dest", outputFile)
 		return err
 	}
 	return nil